Trying to use D3 for the axis
