+++ src/ensembl/stories/entity-viewer/index.ts
@@ -1,0 +1,1 @@
+import './feature-length-axis/FeatureLengthAxis.story';
