+++ src/ensembl/stories/index.tsx
@@ -4,3 +4,4 @@ import './design-primitives/index.ts';
 import './shared-components/index.ts';
 import './static-images/index.ts';
 import './genome-browser/index.ts';
+import './entity-viewer/index.ts';
